requirements.txt
pydantic>=2.5.0
pyyaml>=6.0.0
pytest>=8.0.0
